requirements.txt
fastapi
uvicorn
mcp
openai
anthropic
google-generativeai
arxiv
PyPDF2
requests
python-dotenv
pydantic
asyncio-mqtt
websockets
click
rich
